The last fixes
