+++ cs/Markdown/Markdown/Program.cs
@@ -1,16 +1,10 @@
-﻿using System;
-
-
-namespace Markdown
+﻿namespace Markdown
 {
     class Program
     {
         static void Main(string[] args)
         {
-            var str = "_a __abc__";
-            var result = Md.Render(str);
 
-            Console.WriteLine(result);
         }
     }
 }
